pipeline/requirements.txt
# Pipeline — full deps for scraping, embedding, LLM analysis, email
# Server (pipeline microservice)
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.34.0,<1.0

# Core
python-dotenv>=1.0.1
pyyaml>=6.0.2
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
httpx>=0.28.1,<1.0.0
tenacity>=8.2.0,<9.0.0

# LLM
openai>=1.58.0,<2.0.0
langfuse>=2.0.0,<3.0.0

# Embeddings (torch installed separately as CPU-only in CI)
# Pin transformers <4.48 — PyTorch 2.2.2 is the last build for Intel Macs,
# and transformers >=4.48 requires PyTorch >=2.4.
sentence-transformers>=3.3.1
transformers>=4.41.0,<4.48.0

# Scraping
python-jobspy>=1.1.75
beautifulsoup4>=4.12.3
lxml>=5.3.0

# Database
asyncpg>=0.30.0

# Email verification
dnspython>=2.7.0

# Email sending
aiosmtplib>=3.0.0

# Error monitoring
sentry-sdk[fastapi]>=2.19.0,<3.0

# Testing & Linting
pytest>=8.0.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
